+++ src/graph_view.rs
@@ -2,6 +2,7 @@ use std::borrow::Borrow;
 
 use derive_more::{From, Into};
 use derive_where::derive_where;
+use relrc::RelRcGraph;
 use serde::{Deserialize, Serialize};
 
 use crate::{
@@ -16,7 +17,7 @@ use crate::{
     serialize = "G: Serialize, G::Node: Serialize, G::PortLabel: Serialize, G::Edge: Serialize",
     deserialize = "G: Deserialize<'de>, G::Node: Deserialize<'de>, G::PortLabel: Deserialize<'de>, G::Edge: Deserialize<'de>"
 ))]
-pub struct GraphView<G: Graph>(relrc::GraphView<PortDiffData<G>, EdgeData<G>>);
+pub struct PortDiffGraph<G: Graph>(RelRcGraph<PortDiffData<G>, EdgeData<G>>);
 
 /// A handle to a node in a graph view.
 #[derive(From, Into)]
@@ -30,14 +31,24 @@ impl<'a, G: Graph> From<&'a PortDiff<G>> for NodeId<G> {
     }
 }
 
-impl<G: Graph> GraphView<G> {
+impl<G: Graph> PortDiffGraph<G> {
     pub fn all_nodes(&self) -> impl Iterator<Item = NodeId<G>> + '_ {
         self.0.all_nodes().iter().copied().map(|n| n.into())
     }
 
     pub fn from_sinks(sinks: impl IntoIterator<Item = PortDiff<G>>) -> Self {
-        Self(relrc::GraphView::from_sinks(
+        Self(RelRcGraph::from_sinks(
             sinks.into_iter().map(|n| n.data).collect(),
+        ))
+    }
+
+    pub fn from_sinks_while(
+        sinks: impl IntoIterator<Item = PortDiff<G>>,
+        predicate: impl Fn(&PortDiff<G>) -> bool,
+    ) -> Self {
+        Self(RelRcGraph::from_sinks_while(
+            sinks.into_iter().map(|n| n.data).collect(),
+            |n| predicate(&PortDiff { data: n.clone() }),
         ))
     }
 
@@ -53,17 +64,17 @@ impl<G: Graph> GraphView<G> {
         self.0.merge(other.0);
     }
 
-    pub fn inner(&self) -> &relrc::GraphView<PortDiffData<G>, EdgeData<G>> {
+    pub fn inner(&self) -> &RelRcGraph<PortDiffData<G>, EdgeData<G>> {
         &self.0
     }
 
     pub fn lowest_common_ancestors(graphs: &[Self]) -> impl Iterator<Item = NodeId<G>> + '_ {
-        relrc::GraphView::lowest_common_ancestors(graphs).map(|n| n.into())
+        RelRcGraph::lowest_common_ancestors(graphs).map(|n| n.into())
     }
 }
 
-impl<G: Graph> Borrow<relrc::GraphView<PortDiffData<G>, EdgeData<G>>> for GraphView<G> {
-    fn borrow(&self) -> &relrc::GraphView<PortDiffData<G>, EdgeData<G>> {
+impl<G: Graph> Borrow<RelRcGraph<PortDiffData<G>, EdgeData<G>>> for PortDiffGraph<G> {
+    fn borrow(&self) -> &RelRcGraph<PortDiffData<G>, EdgeData<G>> {
         &self.0
     }
 }